remove cborpretty.c from cmake; fix bug in metrics and defender test
