+++ modules/libraries/3rdparty/CMakeLists.txt
@@ -49,7 +49,13 @@ target_include_directories(
 
 # tinycbor - CBOR implementation
 afr_3rdparty_module(tinycbor)
-afr_glob_files(tinycbor_src DIRECTORY "${AFR_3RDPARTY_DIR}/tinycbor" RECURSE)
+set(tinycbor_src
+    "${AFR_3RDPARTY_DIR}/tinycbor/cborencoder.c"
+    "${AFR_3RDPARTY_DIR}/tinycbor/cborencoder_close_container_checked.c"
+    "${AFR_3RDPARTY_DIR}/tinycbor/cborerrorstrings.c"
+    "${AFR_3RDPARTY_DIR}/tinycbor/cborparser.c"
+    "${AFR_3RDPARTY_DIR}/tinycbor/cborparser_dup_string.c"
+)
 target_sources(
     3rdparty::tinycbor INTERFACE
     ${tinycbor_src}
